move routes to storage constats, 404 page as static
